Add automatic creation of missing load patterns in target etabs file
Whatever load pattern/load case selected by the user, is defined only in the source file and not
in the target file, such load pattern/load case gets added in the target etabs file.
This happens only if the considered load pattern/case has been selected by the user for the transfer
of joints reactions from one model to the other.